pyproject.toml
[project]
name = "communicator-interface"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "customtkinter",
    "matplotlib",
    "pyserial",
    "datetime",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["helpers"]